Checkstyle fixes

+++ jmh/run.sh
@@ -2,18 +2,23 @@
 BASEDIR=$(dirname $0)
 
 
-
 echo "Building RoaringBitmap jar"
 rm -f $BASEDIR/../target/RoaringBitmap*.jar
 mvn -f $BASEDIR/../pom.xml clean install -DskipTests -Dgpg.skip=true
+
+[[ $? -eq 0 ]] || exit
 
 echo "Building Real Roaring Dataset jar"
 rm -f $BASEDIR/../real-roaring-dataset/target/real-roaring-dataset*.jar
 mvn -f $BASEDIR/../real-roaring-dataset/pom.xml clean install
 
+[[ $? -eq 0 ]] || exit
+
 echo "Building benchmarks jar"
 rm -f  $BASEDIR/target/benchmarks.jar
 mvn -f $BASEDIR/pom.xml clean install -Dtest=*$1* -DfailIfNoTests=false
+
+[[ $? -eq 0 ]] || exit
 
 echo "Running benchmarks"
 java -jar $BASEDIR/target/benchmarks.jar  true -wi 5 -i 5 -f 1 $1
